requirements.txt
robotframework
py-canoe
tenma-serial
XlsxWriter
